+++ plugin-server/tests/shared/process-event.ts
@@ -2121,31 +2121,6 @@ export const createProcessEventTests = (
         })
     })
 
-    test('handleTimestamp', () => {
-        const timestamps = [
-            '2021-10-29',
-            '2021-10-29 00:00:00',
-            '2021-10-29T00:00:00.000Z',
-            '2021-10-29 00:00:00+00:00',
-            '2021-10-29T00:00:00.000-00:00',
-            '2021-10-29T00:00:00.000',
-            '2021-10-29T00:00:00.000+00:00',
-            '2021-W43-5',
-            '2021-302',
-        ]
-
-        for (const timestamp of timestamps) {
-            const parsedTimestamp = eventsProcessor.handleTimestamp({ timestamp } as PluginEvent, DateTime.now(), null)
-            expect(parsedTimestamp.year).toBe(2021)
-            expect(parsedTimestamp.month).toBe(10)
-            expect(parsedTimestamp.day).toBe(29)
-            expect(parsedTimestamp.hour).toBe(0)
-            expect(parsedTimestamp.minute).toBe(0)
-            expect(parsedTimestamp.second).toBe(0)
-            expect(parsedTimestamp.millisecond).toBe(0)
-        }
-    })
-
     if (database == 'clickhouse') {
         test('groupidentify', async () => {
             await createPerson(hub, team, ['distinct_id1'])
